pyproject.toml
[project]
name = "senior-design"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py==2.3.1",
    "accelerate==1.10.1",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anywidget==0.9.18",
    "asttokens==3.0.0",
    "attrs==25.4.0",
    "bbox-visualizer==0.2.2",
    "bcrypt>=5.0.0",
    "blinker==1.9.0",
    "boto3>=1.40.68",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "cmake==4.1.2",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "cython==3.1.4",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "einops==0.8.1",
    "executing==2.2.1",
    "fairscale==0.4.13",
    "filelock==3.19.1",
    "filetype==1.2.0",
    "fire==0.7.1",
    "flatbuffers==25.9.23",
    "fonttools==4.60.1",
    "fsspec==2025.9.0",
    "ftfy==6.3.1",
    "git-filter-repo==2.47.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "grpcio==1.75.1",
    "huggingface-hub==0.35.3",
    "humanfriendly==10.0",
    "idna==3.7",
    "ipython==9.6.0",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets==8.1.7",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-bbox-widget==0.6.0",
    "jupyterlab-widgets==3.0.15",
    "kiwisolver==1.4.9",
    "markdown==3.9",
    "markupsafe==3.0.3",
    "matplotlib==3.10.6",
    "matplotlib-inline==0.1.7",
    "ml-dtypes==0.5.3",
    "mpmath==1.3.0",
    "narwhals==2.10.0",
    "networkx==3.5",
    "ninja==1.13.0",
    "numpy==2.3.3",
    "onnx==1.19.1",
    "onnx-ir==0.1.12",
    "onnxruntime",
    "onnxscript==0.5.4",
    "open-clip-torch==3.2.0",
    "opencv-python==4.11.0.86",
    "opencv-python-headless==4.10.0.84",
    "packaging==25.0",
    "pandas==2.3.3",
    "parso==0.8.5",
    "peft==0.17.1",
    "pi-heif==1.1.1",
    "pillow==11.3.0",
    "pillow-avif-plugin==1.5.2",
    "pip==25.3",
    "polygraphy==0.49.26",
    "prompt-toolkit==3.0.52",
    "protobuf==6.32.1",
    "psutil==7.1.0",
    "psygnal==0.14.2",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pybboxes==0.1.6",
    "pycocotools==2.0.10",
    "pydantic==2.11.10",
    "pydantic-core==2.33.2",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pylabel==0.1.55",
    "pyparsing==3.2.5",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "regex==2025.9.18",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rf100vl==1.1.0",
    "roboflow==1.2.10",
    "rpds-py==0.28.0",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "setuptools==80.9.0",
    "shapely==2.1.2",
    "six==1.17.0",
    "smmap==5.0.2",
    "stack-data==0.6.3",
    "streamlit[auth]==1.51.0",
    "supervision==0.26.1",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tensorboard==2.20.0",
    "tensorboard-data-server==0.7.2",
    "termcolor==3.1.0",
    "terminaltables==3.1.10",
    "threadpoolctl==3.6.0",
    "timm==1.0.20",
    "tokenizers==0.22.1",
    "toml==0.10.2",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.57.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "watchdog==6.0.0",
    "wcwidth==0.2.14",
    "werkzeug==3.1.3",
    "widgetsnbextension==4.0.14",
    "rfdetr",
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "django-cors-headers>=4.9.0",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.1",
    "authlib>=1.6.5",
]

[tool.uv.sources]
rfdetr = { path = "./RF-DETR-model_modified/rf-detr-modifications", editable = true }
